Renamed conversion script
