+++ test/test_bsf_2_beios.py
@@ -1,5 +1,5 @@
 import unittest
-from scripts.convert_markup import convert_bsf, parse_bsf, BsfInfo
+from scripts.convert_data import convert_bsf, parse_bsf, BsfInfo
 
 
 class TestBsf2Beios(unittest.TestCase):
